#39: add monthly work details endpoint, update details on shift crud operations

+++ backend/src/main/java/ase/meditrack/service/ShiftService.java
@@ -1,21 +1,35 @@
 package ase.meditrack.service;
 
 import ase.meditrack.exception.NotFoundException;
+import ase.meditrack.model.entity.MonthlyPlan;
 import ase.meditrack.model.entity.Shift;
+import ase.meditrack.model.entity.ShiftType;
+import ase.meditrack.repository.MonthlyPlanRepository;
 import ase.meditrack.repository.ShiftRepository;
+import ase.meditrack.repository.ShiftTypeRepository;
+import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
 @Slf4j
 public class ShiftService {
     private final ShiftRepository repository;
+    private final ShiftTypeRepository shiftTypeRepository;
 
-    public ShiftService(ShiftRepository repository) {
+    private final MonthlyWorkDetailsService monthlyWorkDetailsService;
+    private final MonthlyPlanRepository monthlyPlanRepository;
+
+    public ShiftService(ShiftRepository repository, MonthlyWorkDetailsService monthlyWorkDetailsService,
+                        ShiftTypeRepository shiftTypeRepository, MonthlyPlanRepository monthlyPlanRepository) {
         this.repository = repository;
+        this.monthlyWorkDetailsService = monthlyWorkDetailsService;
+        this.shiftTypeRepository = shiftTypeRepository;
+        this.monthlyPlanRepository = monthlyPlanRepository;
     }
 
     /**
@@ -44,8 +58,18 @@ public class ShiftService {
      * @param shift the shift to create
      * @return the created shift
      */
+    @Transactional
     public Shift create(Shift shift) {
-        return repository.save(shift);
+        Optional<ShiftType> type = shiftTypeRepository.findById(shift.getShiftType().getId());
+        Optional<MonthlyPlan> plan = monthlyPlanRepository.findById(shift.getMonthlyPlan().getId());
+        if (type.isEmpty() || plan.isEmpty()) {
+            throw new NotFoundException("Could not find shift type or plan of shift!");
+        }
+        Shift createdShift = repository.save(shift);
+        createdShift.setShiftType(type.get());
+        createdShift.setMonthlyPlan(plan.get());
+        monthlyWorkDetailsService.updateMonthlyWorkDetailsForShift(createdShift, null);
+        return createdShift;
     }
 
     /**
@@ -56,15 +80,22 @@ public class ShiftService {
      */
     public Shift update(Shift shift) {
         Shift dbShift = findById(shift.getId());
+        Optional<ShiftType> oldShiftType = shiftTypeRepository.findById(dbShift.getShiftType().getId());
+        Optional<ShiftType> newShiftType = shiftTypeRepository.findById(shift.getShiftType().getId());
+        Optional<MonthlyPlan> plan = monthlyPlanRepository.findById(shift.getMonthlyPlan().getId());
+        if (newShiftType.isEmpty() || oldShiftType.isEmpty() || plan.isEmpty()) {
+            throw new NotFoundException("Could not find shift type or plan of shift!");
+        }
+        Shift createdShift = repository.save(shift);
+        createdShift.setShiftType(newShiftType.get());
+        createdShift.setMonthlyPlan(plan.get());
+        monthlyWorkDetailsService.updateMonthlyWorkDetailsForShift(createdShift, oldShiftType.get());
 
         if (shift.getDate() != null) {
             dbShift.setDate(shift.getDate());
         }
         if (shift.getMonthlyPlan() != null) {
             dbShift.setMonthlyPlan(shift.getMonthlyPlan());
-        }
-        if (shift.getShiftType() != null) {
-            dbShift.setShiftType(shift.getShiftType());
         }
         if (shift.getUsers() != null) {
             dbShift.setUsers(shift.getUsers());
@@ -76,7 +107,7 @@ public class ShiftService {
             dbShift.setRequestedShiftSwap(shift.getRequestedShiftSwap());
         }
 
-        return repository.save(dbShift);
+        return repository.save(createdShift);
     }
 
     /**
@@ -85,6 +116,18 @@ public class ShiftService {
      * @param id the id of the shift to delete
      */
     public void delete(UUID id) {
+        Optional<Shift> shift = repository.findById(id);
+        if (shift.isEmpty()) {
+            throw new NotFoundException("Could not find shift to delete!");
+        }
+
+        Optional<ShiftType> newShiftType = shiftTypeRepository.findById(shift.get().getShiftType().getId());
+        Optional<MonthlyPlan> plan = monthlyPlanRepository.findById(shift.get().getMonthlyPlan().getId());
+        if (newShiftType.isEmpty() || plan.isEmpty()) {
+            throw new NotFoundException("Could not find shift type or plan of shift!");
+        }
+
+        monthlyWorkDetailsService.updateMonthlyWorkDetailsForDeletedShift(shift.get());
         repository.deleteById(id);
     }
 }
